Refactor ReadBuffer::parse_with to use a Parser trait.
This is needed to help support decoding unknown or custom blocks.

We need it to be possible to return a reference to the current
parser state, along with the next item from the parser.

To do this we have to split up the parsing into two steps:

1. Getting the next item.
2. Updating the parser state.

The first takes an immutable reference to the parser, and may be looped.
The second takes a mutable reference to the parser, and is done only once.

After both steps are done, we take an immutable reference to the parser
state and return it along with the item from the parser.